+++ src/main/java/org/kinscript/Academy/dominio/dto/GradosDto.java
@@ -3,7 +3,7 @@ package org.kinscript.Academy.dominio.dto;
 import jakarta.validation.constraints.NotBlank;
 
 public record GradosDto(
-        Integer idGrado,
+        Long idGrado,
         @NotBlank(message = "El nombre del grado es obligatorio")
         String nombreGrado
 ) {
