+++ src/main/java/org/kinscript/Academy/dominio/dto/NotasDto.java
@@ -7,13 +7,13 @@ import jakarta.validation.constraints.DecimalMax;
 import java.math.BigDecimal;
 
 public record NotasDto(
-        Integer idNota,
+        Long idNota,
 
         @NotNull(message = "El ID del alumno es obligatorio")
-        Integer idAlumno,
+        Long idAlumno,
 
         @NotNull(message = "El ID del curso es obligatorio")
-        Integer idCurso,
+        Long idCurso,
 
         @DecimalMin(value = "0.00", message = "La nota no puede ser menor a 0")
         @DecimalMax(value = "100.00", message = "La nota no puede ser mayor a 100")
