+++ src/main/java/org/kinscript/Academy/dominio/dto/SeccionesDto.java
@@ -3,7 +3,7 @@ package org.kinscript.Academy.dominio.dto;
 import jakarta.validation.constraints.NotBlank;
 
 public record SeccionesDto(
-        Integer idSeccion,
+        Long idSeccion,
         @NotBlank(message = "El nombre de la sección es obligatorio")
         String nombreSeccion
 ) {
