+++ src/main/java/org/kinscript/Academy/dominio/exception/AlumnoNotExistsException.java
@@ -1,7 +1,7 @@
 package org.kinscript.Academy.dominio.exception;
 
 public class AlumnoNotExistsException extends RuntimeException {
-    public AlumnoNotExistsException(Integer codigo) {
+    public AlumnoNotExistsException(Long codigo) {
         super("No existe un alumno con el ID: " + codigo);
     }
 }
