+++ src/main/java/org/kinscript/Academy/dominio/exception/NotaExistsException.java
@@ -1,7 +1,7 @@
 package org.kinscript.Academy.dominio.exception;
 
 public class NotaExistsException extends RuntimeException {
-    public NotaExistsException(Integer idAlumno, Integer idCurso) {
+    public NotaExistsException(Long idAlumno, Long idCurso) {
         super("Ya existe un registro de notas para el alumno con ID '" + idAlumno + "' en el curso con ID '" + idCurso + "'.");
     }
 }
