+++ src/main/java/org/kinscript/Academy/dominio/exception/ProfesorAsignacionNotExistsException.java
@@ -1,7 +1,7 @@
 package org.kinscript.Academy.dominio.exception;
 
 public class ProfesorAsignacionNotExistsException extends RuntimeException {
-    public ProfesorAsignacionNotExistsException(Integer idProfesor, Integer idCurso, Integer idGrado, Integer idSeccion) {
+    public ProfesorAsignacionNotExistsException(Long idProfesor, Long idCurso, Long idGrado, Long idSeccion) {
         super("No existe la asignación para el profesor " + idProfesor + ", curso " + idCurso +
                 ", grado " + idGrado + " y sección " + idSeccion + ".");
     }
